+++ temba/settings_common.py
@@ -476,7 +476,12 @@ GROUP_PERMISSIONS = {
     "Customer Support": (
         "auth.user_list",
         "auth.user_update",
+        "campaigns.campaign_read",
+        "channels.channel_configuration",
+        "channels.channel_history",
+        "channels.channel_read",
         "contacts.contact_break_anon",
+        "contacts.contact_read",
         "flows.flow_editor",
         "flows.flow_json",
         "flows.flow_revisions",